Fix download/delete error on uploaded files.
- Since the saved files are now stored in the session,
the merge is not needed.
- Add also a check for unlink files.

+++ phprojekt/library/Phprojekt/Item/Abstract.php
@@ -285,9 +285,9 @@ abstract class Phprojekt_Item_Abstract extends Phprojekt_ActiveRecord_Abstract i
                 $filesField = $this->$field;
                 $files      = explode('||', $filesField);
                 foreach ($files as $file) {
-                    $md5Name = substr($file, 0, strpos($file, '|'));
+                    $md5Name          = substr($file, 0, strpos($file, '|'));
                     $fileAbsolutePath = Phprojekt::getInstance()->getConfig()->uploadpath . $md5Name;
-                    if (file_exists($fileAbsolutePath)) {
+                    if (!empty($md5Name) && file_exists($fileAbsolutePath)) {
                         unlink($fileAbsolutePath);
                     }
                 }
